requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
gspread==6.1.2
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.5.4
python-dotenv==1.0.1
tqdm==4.66.4
rapidfuzz==3.9.3